UC-12 refactor: Post slice with segments
