--- src/entities/PolicySwitcher/model/types.ts
@@ -1,3 +1,0 @@
-export type PolicySwitcherProps = {
-  variant: 'policy' | 'terms'
-}
