+++ src/entities/Post/ui/index.ts
@@ -5,5 +5,5 @@ export { PostLower } from './post-lower/PostLower'
 export { AvatarImageLink } from './AvatarImageLink'
 export { PostIconButton } from './PostIconButton'
 
-export { truncateDescription } from '../lib'
+export { truncateDescription } from '../lib/truncateDescription'
 export * from '../model/types'
